+++ Changelog.md
@@ -1,5 +1,10 @@
 # Rdrillog Changelog
 
+## Version 2.1.0
+- Adding sampling option -> It permit to choose the period used for sampling in graphs
+
+- When --erase false was used, the output file was erased. -> Bug fixed 
+
 ## Version 2.0.0
 - Adding graph feature: BarChart, GaugeChart, TrendChart
 - Adding the possibility to enable or disable graph feature